Create LICENSE
I think that should be enough. It is the exact same as "Unlicensed" or "No-License" but with the addition of reproducing the code or re-compiling it WITHOUT being able to distribute it as a new product. That is to allow all modders and mods a happy time.

Whatever they do with their mods are up to them.